Switch CI to Ubuntu 20.04, add clang-10 build (#2100)
* refs #2074
* Adds a ubuntu/clang-10 build in CI but sticks with clang-9 as the default compiler. Mac is unchanged.
* Adds `cmake -DHYRISE_RELAXED_BUILD=On ..`, which disables most warnings.
* I verified that all scripts still work for Ubuntu 19.04, so while it is not our preferred system anymore, we don't break anything.

+++ src/test/operators/join_test_runner.cpp
@@ -233,11 +233,12 @@ class JoinTestRunner : public BaseTestWithParam<JoinTestConfiguration> {
           const auto chunk_count = indexed_input.table_size == 0
                                        ? uint32_t{0}
                                        : static_cast<uint32_t>(std::ceil(static_cast<float>(indexed_input.table_size) /
-                                                                         indexed_input.chunk_size));
+                                                                         static_cast<float>(indexed_input.chunk_size)));
 
           auto indexed_chunk_ranges = std::set<ChunkRange>{};
           for (const auto indexed_share : indexed_segment_shares) {
-            indexed_chunk_ranges.insert(ChunkRange{0, static_cast<uint32_t>(std::ceil(indexed_share * chunk_count))});
+            indexed_chunk_ranges.insert(
+                ChunkRange{0, static_cast<uint32_t>(std::ceil(indexed_share * static_cast<float>(chunk_count)))});
           }
 
           for (const auto& indexed_chunk_range : indexed_chunk_ranges) {
